moved AppLockScreen from dialogs to package lockscreen
