+++ app/src/main/java/com/babylon/wallet/android/presentation/lockscreen/AppLockActivity.kt
@@ -1,6 +1,5 @@
-package com.babylon.wallet.android.presentation.dialogs.lock
+package com.babylon.wallet.android.presentation.lockscreen
 
-import android.app.Activity
 import android.os.Build
 import android.os.Bundle
 import androidx.activity.compose.setContent
@@ -13,11 +12,12 @@ class AppLockActivity : FragmentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         setContent {
             RadixWalletTheme {
                 AppLockScreen(onUnlock = {
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-                        overrideActivityTransition(Activity.OVERRIDE_TRANSITION_CLOSE, 0, 0)
+                        overrideActivityTransition(OVERRIDE_TRANSITION_CLOSE, 0, 0)
                     } else {
                         overridePendingTransition(0, 0)
                     }
