+++ app/src/main/java/com/babylon/wallet/android/presentation/lockscreen/AppLockViewModel.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.dialogs.lock
+package com.babylon.wallet.android.presentation.lockscreen
 
 import com.babylon.wallet.android.AppLockStateProvider
 import com.babylon.wallet.android.presentation.common.StateViewModel
@@ -14,13 +14,11 @@ class AppLockViewModel @Inject constructor(
     private val deviceCapabilityHelper: DeviceCapabilityHelper,
 ) : StateViewModel<AppLockViewModel.State>() {
 
-    fun onUnlock() {
-        appLockStateProvider.unlockApp()
+    override fun initialState(): State {
+        return State(
+            isDeviceSecure = deviceCapabilityHelper.isDeviceSecure
+        )
     }
-
-    data class State(
-        val isDeviceSecure: Boolean
-    ) : UiState
 
     fun onResumed() {
         _state.update {
@@ -30,9 +28,11 @@ class AppLockViewModel @Inject constructor(
         }
     }
 
-    override fun initialState(): State {
-        return State(
-            isDeviceSecure = deviceCapabilityHelper.isDeviceSecure
-        )
+    fun onUnlock() {
+        appLockStateProvider.unlockApp()
     }
+
+    data class State(
+        val isDeviceSecure: Boolean
+    ) : UiState
 }
